feat(#233): extract scheduler engine into pure lib (Phase 2 of Resource Optimiser)
